update
update

+++ CubeTesten/VS Project/ConsoleApplication1/ConsoleApplication1/Object.cpp
@@ -37,7 +37,7 @@ uint16_t Object::getLposX()
 	return this->LposX;
 }
 
-uint8_t Object::getSpritenr()
+uint16_t Object::getSpritenr()
 {
 	return this->spritenr;
 }
