+++ CubeTesten/VS Project/ConsoleApplication1/ConsoleApplication1/Object.h
@@ -13,7 +13,7 @@ public:
 	void changeLposY(uint16_t);
 	uint16_t getLposX();
 	uint16_t getLposY();
-	uint8_t getSpritenr();
+	uint16_t getSpritenr();
 	uint16_t getMiddleX();
 	uint16_t getMiddleY();
 	void setNext(Object *);
@@ -22,17 +22,17 @@ public:
 	Object * getPrev();
 
 public:
-	uint8_t UniqueIDWhenactive = 0;
+	uint16_t UniqueIDWhenactive = 0;
 	Object * Next = nullptr;
 	Object * Prev = nullptr;
-	uint8_t spritenr = 0;
+	uint16_t spritenr = 0;
 	uint16_t Xlength = 0;//X length
 	uint16_t Ylength = 0;//Y length
 	uint16_t LposX = 0;
 	uint16_t LposY = 0;
 	uint16_t Radius = 0;
-	uint8_t speed = 0;
-	uint8_t gravity = 0;
+	uint16_t speed = 0;
+	uint16_t gravity = 0;
 
 	friend class Game;
 	friend class Platform;
